% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_wind_wave.R
\name{build_wind_wave}
\alias{build_wind_wave}
\title{Generate and Propagate a Wind Wave Over Variable Bathymetry}
\usage{
build_wind_wave(fetch, depths, distances, wind_speed)
}
\arguments{
\item{fetch}{Numeric. Total wind fetch distance (in meters).}

\item{depths}{Numeric vector Water depths (in meters) at each segment; must
be one element longer than \code{distances}.}

\item{distances}{Numeric vector. Horizontal distances (in meters) between successive
depth points. Must be one element shorter than \code{depths}.}

\item{wind_speed}{Numeric. Wind speed blowing from direction (in m/s).}
}
\value{
A data frame with the following columns:
\describe{
\item{wave_height_final}{Final wave height at the end of the fetch (in meters).}
\item{WEI}{Wave Energy Index}
\item{wave_period}{Estimated wave period (in seconds).}
\item{wave_number}{Average wave number (radians per meter).}
\item{celerity_final}{Final wave celerity (phase speed, in m/s).}
\item{nnumber_final}{Final group velocity coefficient.}
}
}
\description{
This function builds a wind-wave over input bathymetry and distances
returning an estimate of final wave height and related wave characteristics
(e.g., wave energy index, wave period, wave number, celerity)
}
\details{
Wave growth is calculated segment-by-segment, adjusting for
shoaling, wave breaking, and changing depth. Deep and shallow water regimes
are handled using different formulas. The function uses a Newton-Raphson
method to compute wave number.
}
\examples{
fetch <- 90
depths <- c(1, 2, 3, 5, 5)
distances <- c(25, 25, 25, 15)
wind_speed <- 10
wind_proportion <- 0.25
result <- build_wind_wave(fetch, depths, distances, wind_speed)

}
