% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassDegrees_to_radianDegrees.R
\name{compassDegrees_to_radianDegrees}
\alias{compassDegrees_to_radianDegrees}
\title{Convert compass degrees (direction) to radian degrees}
\usage{
compassDegrees_to_radianDegrees(compassDegrees)
}
\arguments{
\item{compassDegrees}{compass direction or heading in degrees}
}
\value{
a numerical vector
}
\description{
Convert compass degrees (direction) to radian degrees
}
\examples{
# North on a compass is 0 deg
compassDegrees_to_radianDegrees(0)

# East on a compass is 90 deg
compassDegrees_to_radianDegrees(90)

# South on a compass is 180 deg
compassDegrees_to_radianDegrees(180)

# West on a compass is 270 deg
compassDegrees_to_radianDegrees(270)

# North on a compass is ALSO 360 deg
compassDegrees_to_radianDegrees(360)
}
\keyword{internal}
