% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bounding_box.R
\name{create_bounding_box}
\alias{create_bounding_box}
\title{Create a bounding box around a point with an expansion distance}
\usage{
create_bounding_box(site_point, expansion_dist)
}
\arguments{
\item{site_point}{a SpatVector (or coercible to a SpatVector)}

\item{expansion_dist}{a vector of length 1 or 2 to add to the limits of the bounding box. If length 1, the value is used for both x and y expansion}
}
\value{
a SpatExtent object
}
\description{
Create a bounding box around a point with an expansion distance
}
\examples{
site_point <- terra::vect(cbind(x = 406825, y = 4372214), crs = "EPSG:26918")
create_bounding_box(site_point, 1000)
}
