% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NOAA_CUDEM.R
\name{download_NOAA_CUDEM}
\alias{download_NOAA_CUDEM}
\title{Download NOAA NCEI CUDEM (Continuously Updated DEM)}
\usage{
download_NOAA_CUDEM(
  tile_index,
  lon,
  lat,
  radius_m,
  dest_dir = "NOAA_CUDEM",
  overwrite = FALSE
)
}
\arguments{
\item{lon}{Longitude of the center point (in decimal degrees, WGS84).}

\item{lat}{Latitude of the center point (in decimal degrees, WGS84).}

\item{radius_m}{Radius around the point, in meters, within which to download intersecting tiles.}

\item{dest_dir}{Directory to save downloaded files. Created if it does not exist. Default is \code{"NOAA_CUDEM"}.}

\item{overwrite}{Logical. If \code{TRUE}, re-download files even if they already exist locally. Default is \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} if downloads completed, or \code{NULL} if no tiles were within the radius.

Download Tiles Near a Point Within a Radius

...
}
\description{
Downloads Bathymetric-Topographic rasters files from the NOAA NCEI
Continuously Updated Digital Elevation Model (CUDEM) – Ninth Arc-Second
Resolution. The tiles are selected based on proximity to a user-defined point
(longitude/latitude) and radius (in meters), and intersecting tiles are
downloaded via their individual URLs stored in a built-in tile index.
}
\details{
A tile index is used to identify proximate raster files. This file is included in this
package and can be downloaded separately from
\href{https://coast.noaa.gov/htdata/raster2/elevation/NCEI_ninth_Topobathy_2014_8483/#:~:text=com/dem/NCEI_ninth_Topobathy_2014_8483/.-,Meta\%20Info,-Tile\%20Index\%3A}{coast.noaa.gov}
}
\section{Data Source}{
 The dataset is provided by the \strong{NOAA National Centers
for Environmental Information (NCEI)}:
\itemize{
\item NOAA NCEI CUDEM Info/Metadata/Download Page:
\url{https://coast.noaa.gov/htdata/raster2/elevation/NCEI_ninth_Topobathy_2014_8483/}
}
}

