% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wind_data.R
\name{get_wind_data}
\alias{get_wind_data}
\title{Retrieve and clean wind data for WEMo}
\usage{
get_wind_data(site_point, years, which_station = "ask")
}
\arguments{
\item{site_point}{A spatial point object (either \code{sf} or coercible to \code{sf})
indicating the target location.}

\item{years}{A vector of years (e.g., \code{2002:2024}) for which to download wind
data.}

\item{which_station}{Either:
\itemize{
\item \code{"ask"}: interactively choose from 5 closest stations;
\item an integer (1-5): pick the nth closest station;
\item a string station code in the form "USAF-WBAN".
}}
}
\value{
A data frame with cleaned wind direction and wind speed, along with
station code, timestamp, and date components (year, month, day).
}
\description{
This function is a wrapper around a workflow to find and download wind data
using the \link[worldmet:import_ghcn_stations]{import_ghcn_stations()} and
\link[worldmet:import_ghcn_hourly]{import_ghcn_hourly()} functions from the
\strong{worldmet} package. It finds up to 5 nearby NOAA GHCN weather stations based
on a geographic point and downloads hourly wind data.
}
\details{
This function utilizes data from \strong{Global Historical Climatology
Network (GHCN)} dataset. GHCN aggregates hourly meteorological
observations from numerous fixed, land-based stations. maintained by NOAA,
the U.S. Air Force, and many other meteorological agencies (Met Services)
around the world.
}
\examples{
# Create a POINT geometry from coordinates and cast explicitly
site_point <- sf::st_sf(
  geometry = sf::st_sfc(sf::st_point(c(-76.67587, 34.71413))),
  crs = 4326
  )

# 1. Prompt user to select a station interactively
\dontrun{
wind_data_ask <- get_wind_data(
  site_point,
  years = 2022:2023,
  which_station = "ask"
 )
}

# 2. Automatically use the 2nd closest station
\dontrun{
wind_data_index <- get_wind_data(
  site_point,
  years = 2022:2023,
  which_station = 2
 )
}

# 3. Manually specify a station code
\dontrun{
# this way you don't need a site_point
wind_data_manual <- get_wind_data(
  site_point = NULL,
  years = 2022:2023,
  which_station = "USW00093765"
 )
}

}
\references{
For more information on the GHCN and to view an interactive
map of stations, see
https://www.ncei.noaa.gov/products/global-historical-climatology-network-hourly
}
