% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PI_points}
\alias{PI_points}
\title{Example Site Points for WEMo}
\format{
\subsection{\code{PI_points}}{

A Simple feature collection with 3 features and 2 fields
\describe{
\item{FID}{feature ID number}
\item{site}{site number}
...
}
}
}
\usage{
PI_points
}
\description{
Point data to be used in examples from wemo package
}
\keyword{datasets}
