% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PI_shoreline}
\alias{PI_shoreline}
\title{Example Shoreline Polygon for WEMo}
\format{
\subsection{\code{PI_shoreline}}{

A Simple feature collection with 3 features and 2 fields
\describe{
\item{FID}{feature ID number}
\item{site}{site number}
...
}
}
}
\usage{
PI_shoreline
}
\description{
A polygon enclosing land areas around Pivers Island NC. The boundary was
drawn at 0.552 m NAVD88 which is MHHW for the area
}
\keyword{datasets}
