% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wind_data.R
\name{summarize_wind_data}
\alias{summarize_wind_data}
\title{Summarize wind data by direction and intensity}
\usage{
summarize_wind_data(
  wind_data,
  wind_percentile,
  directions = NULL,
  wind_speed_na.rm = TRUE
)
}
\arguments{
\item{wind_data}{A data frame containing at least the columns
\code{wind_direction} and \code{wind_speed}}

\item{wind_percentile}{either 'mean' or a numeric between 0 and 1 specifying
the percentile of \code{wind_speed} to calculate}

\item{directions}{numeric vector of wind directions to snap input
\code{wind_direction} to. This is useful if input \code{wind_data$wind_data} has
anomolous \code{wind_direction} readings}

\item{wind_speed_na.rm}{logical. indicating if \code{wind_speed} of \code{NA} indicates
bad data (\code{TRUE}, the default) or calm wind (\code{FALSE}). If \code{FALSE}, the \code{NA}
value will increase the \code{n} and \code{proportion} values but not effect the \code{speed}
value in the returned object}
}
\value{
A tibble summarizing wind by direction and intensity, with columns:
\describe{
\item{direction}{Wind direction (degrees, with 360 converted to 0)}
\item{n}{Count of observations in each direction}
\item{proportion}{Percentage of total observations in each direction}
\item{speed}{Mean or percentile wind speed for each direction}
}
}
\description{
Generates a summary data.frame suitable for passing to \code{\link[=wind_rose]{wind_rose()}} for
plotting or \code{\link[=WEMo]{WEMo()}} for modeling. Input data should be a log of wind readings
and must have columns \code{wind_direction} and \code{wind_speed}. Data generated from
\code{\link[=get_wind_data]{get_wind_data()}} is ready to be used as input. Input \code{wind_direction}
values are coerced into values 0-359.9. The function can further coerce or snap input
\code{wind_direction}s to a user supplied vector of suitable directions
(\code{directions}). This is often necessary with input data from public data sets
that have occasional anomalous readings.  If \code{directions} is not supplied the
output is based on all unique \code{wind_direction}.
}
\examples{
set.seed(1) #for reproduceability
# generate some example data
wind_data <- data.frame(
  wind_direction = sample(c(90, 180, 270, 360), size = 25, replace = TRUE),
  wind_speed = runif(25, 0, 25)
)

# Summarize using the 95th percentile for wind speed
summarize_wind_data(wind_data, wind_percentile = 0.95)

# Summarize using the mean wind speed
summarize_wind_data(wind_data, wind_percentile = "mean")

# create some dummy data with NAs in the speed and direction columns
wind_data <- data.frame(
  wind_direction = c(runif(n = 5, min = 0, max = 315),rep(NA, 10)),
  wind_speed = c(runif(5, 0, 25), rep(NA, 5), rep(0, 5))
)

# remove the NAs in the speed column. perhaps these are data collection errors
summarize_wind_data(wind_data, wind_percentile = 'mean',
  directions = c(90, 180, 270, 360), wind_speed_na.rm = TRUE)

# keep the NAs in the speed column. perhaps these represent calm data
# (though normally 0 would indicate calm)
summarize_wind_data(wind_data, wind_percentile = 'mean',
  directions = c(90, 180, 270, 360), wind_speed_na.rm = FALSE)

}
