% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_depths.R
\name{update_depths}
\alias{update_depths}
\title{Update Depths in a WEMo input}
\usage{
update_depths(wemo_input, depth_diff)
}
\arguments{
\item{wemo_input}{\code{sf} object with column \code{depths} which is a list of vectors}

\item{depth_diff}{numeric. indicates the amount to increase (or decrease for negative values)}
}
\value{
\code{sf} object ready for processing in \code{wemo()}
}
\description{
Update the depth values in a object created by \code{prepare_wemo_inputs()}.
}
