% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEMo-package.R
\docType{package}
\name{WEMo-package}
\alias{WEMo}
\alias{WEMo-package}
\title{WEMo: Wave Energy Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

WEMo is a model for quantifying the wave energy environment at a given location based on wind-generated waves. It uses wind history, bathymetry, and shoreline geometry to estimate wave conditions. This package provides tools to calculate fetch, determine effective fetch, extract bathymetry along vector rays, and model wind-wave properties. It also includes helper functions to locate, download, and prepare inputs such as wind data and shoreline polygons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSE-NCCOS-NOAA/WEMo}
  \item \url{https://mse-nccos-noaa.github.io/WEMo/}
  \item Report bugs at \url{https://github.com/MSE-NCCOS-NOAA/WEMo/issues}
}

}
\author{
\strong{Maintainer}: Quentin Walker \email{quetin.walker@noaa.gov} (\href{https://orcid.org/0000-0002-1375-822X}{ORCID})

}
\keyword{internal}
