% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid_points.R
\name{generate_grid_points}
\alias{generate_grid_points}
\title{Generate a Grid of Points Over a Spatial Extent}
\usage{
generate_grid_points(site_point, expansion_dist, resolution)
}
\arguments{
\item{site_point}{A \code{SpatVector}. The spatial object whose extent will define
the bounds of the grid.}

\item{expansion_dist}{Numeric vector of length 1 or 2. Distance (in map
units) to expand the bounding box in the x and y directions. If a single
value is provided, it is used for both directions. To prevent expansion and
use the exact bounding box of \code{site_point}, set this to \code{0}.}

\item{resolution}{Numeric vector of length 1 or 2. The spacing between grid
points in the x and y directions (in map units). If a single value is
provided, it is used for both directions.}
}
\value{
A \code{SpatVector} of points spaced regularly over the (expanded) extent
of \code{site_point}.
}
\description{
Creates a regular grid of points covering the (optionally expanded) extent of
a \code{SpatVector} object at a specified resolution. Returns the grid as a
\code{SpatVector} of point geometry.
}
\examples{
# single point ----------------------------
site_pt <- terra::vect(
  data.frame(x = -76.67587, y = 34.71413),
  geom = c("x", "y"),
  crs = "EPSG:4326"
  )

# Create a grid of points covering an area expanded by 0.001째 (x) and 0.0002째
# (y) # around the input point, with grid points spaced every 0.0001째 (x) and
# 0.00005째 (y)
grid_pts <- generate_grid_points(
  site_point = site_pt,
  expansion_dist = c(0.001, 0.0002),
  resolution = c(0.0001, 0.00005)
 )

# terra::plot(grid_pts)

# multiple points ---------------------------
site_pts <- terra::vect(
  data.frame(
    x = c(346536.4, 346986.4, 346986.4),
    y = c(3842620,  3842670,  3843270)),
  geom = c("x", "y"),
  crs = "EPSG:26918")

# create a grid of points covering an area expanded by 1000 map units
# (meters) with points spaced every 50 map units (meters)
grid_pts <- generate_grid_points(
  site_point = site_pts,
  expansion_dist = 1000,
  resolution = 50
 )

# terra::plot(grid_pts)

}
