% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PI_wind_data}
\alias{PI_wind_data}
\title{Example Wind Data for WEMo examples}
\format{
\subsection{\code{PI_wind_data}}{

A tibble with 184,947 rows and 73 columns
\describe{
\item{code}{the ISD station code}
\item{time}{datetime when the observation was made}
\item{year}{year when the observation was made}
\item{month}{month when the observation was made}
\item{day}{day when the observation was made}
\item{wind_direction}{direction from which the wind blows. degrees off north}
\item{speed}{speed of the wind. meters per second}
...
}
}
}
\source{
created by:
get_wind_data(
site_point = NULL,
years = 2023:2024,
which_station = "723090-13754"
)
}
\usage{
PI_wind_data
}
\description{
Wind history data from
}
\keyword{datasets}
