% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wind_data.R
\name{summarize_wind_data}
\alias{summarize_wind_data}
\title{Summarize wind data by direction and intensity}
\usage{
summarize_wind_data(wind_data, wind_percentile)
}
\arguments{
\item{wind_data}{A data frame containing at least the columns \code{wind_direction} and \code{wind_speed}}

\item{wind_percentile}{either 'mean' or a numeric between 0 and 1 specifying the percentile of wind speed to calculate}
}
\value{
A data frame summarizing wind by direction, with columns:
\describe{
\item{directions}{Wind direction (degrees, with 360 converted to 0)}
\item{n}{Count of observations in each direction}
\item{proportion}{Percentage of total observations in each direction}
\item{speed}{Mean or percentile wind speed for each direction}
}
}
\description{
Summarize wind data by direction and intensity
}
\examples{

wind_data <- data.frame(
  wind_direction = sample(c(90, 180, 270, 360), size = 25, replace = TRUE),
  wind_speed = runif(25, 0, 25)
)

# Summarize using the 95th percentile for wind speed
summarize_wind_data(wind_data, 0.95)

# Summarize using the mean wind speed
summarize_wind_data(wind_data, "mean")

}
