% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEMo-package.R
\docType{package}
\name{WEMo-package}
\alias{WEMo-package}
\title{WEMo: Wave Energy Model}
\description{
WEMo is a model that quantifies the energetic environment of a given point resulting from wind-generated waves. The model takes as inputs a bathymetry, a shoreline polygon, and wind history data. This packages contains functions to calculate the fetch, determine the effective fetch, grab the bathymetery along a vector ray and then build a wind-wave.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/QAWalker/WEMo}
  \item Report bugs at \url{https://github.com/QAWalker/WEMo/issues}
}

}
\author{
\strong{Maintainer}: Quentin Walker \email{quetin.walker@noaa.gov} (\href{https://orcid.org/0000-0002-1375-822X}{ORCID})

}
\keyword{internal}
