% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wemo.R
\name{wemo}
\alias{wemo}
\title{Run the WEMo Wave Energy Model}
\usage{
wemo(fetch)
}
\arguments{
\item{fetch}{A data frame or \code{sf} object containing fetch geometry and
required variables: \code{efetch}, \code{depths}, \code{distances}, \code{speed}, \code{proportion}.
Expects the output from \code{\link[=prepare_wemo_inputs]{prepare_wemo_inputs()}}. Each row
corresponds to a fetch ray.}
}
\value{
A list of two elements:
\describe{
\item{\code{wemo_details}}{the original \code{fetch} object with the following stats about each wave arriving at the site from the direction of the fetch ray added:
\itemize{
\item \code{wave_height_final} - Final wave height at the end of the fetch (m)
\item \code{WEI} - Wave Energy Index
\item \code{wave_period} - Estimated wave period (s)
\item \code{wave_number} - Average wave number (rad/m)
\item \code{celerity_final} - Final wave phase speed (m/s)
\item \code{nnumber_final} - Final group velocity coefficient
}}
\item{\code{wemo_final}}{sf object with all the variables from \code{site_point} in addition to the following:\itemize{
\item \code{RWE} - Total relative wave exposure at the site
\item \code{avg_wave_height} - Mean wave height across all fetch rays (m)
\item \code{max_wave_height} - Maximum wave height across all fetch rays (m)
\item \code{direction_of_max_wave} - Direction(s) of maximum wave height (Degrees from North)
}}
}
}
\description{
\code{wemo()} Combines fetch, bathymetry, and wind data to estimate
wave height and energy metrics at each site using WEMo (Wave Energy Model)
from wind-waves.

First, the function builds a wind-wave over input bathymetry and returns an
estimate of final wave height and related wave characteristics (e.g., wave
energy index, wave period, wave number, celerity).

Then, the function summarizes the waves from every direction for each site
and computes the summary statistic \code{RWE}.
}
\details{
This function builds and propagates waves arriving at the site from
each fetch ray segment-by-segment, adjusting for shoaling, wave breaking,
and changing depth. Deep and shallow water regimes are handled using
different formulas. The function uses a Newton-Raphson method to compute
wave number. The values in the \code{wemo_final} return are calculated from
summarizing (mean and max wave height) of all fetch rays. Relative Wave
Energy (\code{RWE}) is calculated by multiplying the \code{REI} from each wave by the
proportion of time that wind blows from the direction of the ray each ray
and summing for all fetch rays
}
\examples{
\dontrun{
# Example assumes fetch object includes efetch, depths, distances, speed, and direction
result <- wemo(fetch = fetch, site_points = site_pts)
details <- result$wemo_details
summary <- result$wemo_output
}
}
\seealso{
\code{\link[=prepare_wemo_inputs]{prepare_wemo_inputs()}}
}
